+++ models/person.py
@@ -1,0 +1,18 @@
+class Person:
+
+    def __init__(self, firstName, lastName, email):
+        self.firstName = firstName
+        self.lastName = lastName
+        self.email = email
+
+    def getFirstName(self):
+        return self.firstName
+
+    def getLastName(self):
+        return self.lastName
+
+    def getEmail(self):
+        return self.email
+
+    def printAllDetails(self):
+        print(self.firstName, self.lastName, self.email)
